merge da super integração (falta testes)

See merge request sistema-de_gerenciador_bicicletario/ms-aluguel!20

+++ src/main/java/com/sistema_bicicletario/ms_aluguel/dtos/DevolucaoDTO.java
@@ -12,6 +12,6 @@ public class DevolucaoDTO {
     private LocalDateTime horaInicio;
     private Integer trancaFim;
     private LocalDateTime horaFim;
-    private Integer cobranca;
+    private Long cobranca;
     private Integer ciclista;
 }
